some last changes

+++ app.py
@@ -56,26 +56,22 @@ def get_todos():
         todos = Todo.query.all()
         return todos
 
-@app.route("/update/<int:todo_id>", methods=['GET', 'POST'])
-def update(todo_id):
+@app.route("/change/<int:todo_id>/<int:task>", methods=['GET', 'POST'])
+def change(todo_id, task):
     todos = Todo.query.all()
-    for todo in todos:
-        if(todo.id == todo_id):
-            todo.completed = not todo.completed
-            db.session.commit()
-            break
+    if(task == 0):
+        for todo in todos:
+            if(todo.id == todo_id):
+                todo.completed = not todo.completed
+                db.session.commit()
+                break
+    else:
+        for todo in todos:
+            if(todo.id == todo_id):
+                db.session.delete(todo)
+                db.session.commit()
+                break
     return redirect(url_for("home_route"))
-
-
-@app.route('/delete/<int:todo_id>', methods=['GET', 'POST'])
-def delete_route(todo_id):
-    todos = Todo.query.all()
-    for todo in todos:
-        if(todo.id == todo_id):
-            db.session.delete(todo)
-            db.session.commit()
-            break
-    return redirect(url_for('home_route'))
 
 if __name__ == '__main__':
     app.run(debug=True, host='127.0.0.1', port=5001)
